identity/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
PyJWT==1.7.1